WIP floats for milliseconds

+++ eon_timer/timers/delay_timer.py
@@ -18,25 +18,25 @@ class DelayTimer:
         self.calibrator: Final[Calibrator] = calibrator
         self.second_timer: Final[SecondTimer] = second_timer
 
-    def create(self, target_delay: int, target_second: int, calibration: int) -> list[int]:
+    def create(self, target_delay: int, target_second: int, calibration: float) -> list[float]:
         phases = [
             self.__create_phase1(target_delay, target_second, calibration),
             self.__create_phase2(target_delay, calibration)
         ]
         return phases
 
-    def __create_phase1(self, target_delay: int, target_second: int, calibration: int) -> int:
+    def __create_phase1(self, target_delay: int, target_second: int, calibration: float) -> float:
         return timers.to_minimum_length(
             self.second_timer.create(target_second, calibration)[0] -
             self.calibrator.to_milliseconds(target_delay)
         )
 
-    def __create_phase2(self, target_delay: int, calibration: int) -> int:
+    def __create_phase2(self, target_delay: int, calibration: float) -> float:
         return self.calibrator.to_milliseconds(target_delay) - calibration
 
-    def calibrate(self, target_delay: int, delay_hit: int) -> int:
+    def calibrate(self, target_delay: int, delay_hit: int) -> float:
         delta = self.calibrator.to_milliseconds(delay_hit) - self.calibrator.to_milliseconds(target_delay)
         if abs(delta) <= self.CLOSE_THRESHOLD:
-            return int(self.CLOSE_UPDATE_FACTOR * delta)
+            return self.CLOSE_UPDATE_FACTOR * delta
         else:
-            return int(self.UPDATE_FACTOR * delta)
+            return self.UPDATE_FACTOR * delta
